requirements.txt
streamlit>=1.33
pandas>=2.2
bcrypt==4.1.3
cffi==1.16.0
wheel
setuptools
python-dotenv>=1.0